requirements.txt
fastapi
uvicorn
pydantic
kuberneteswebsockets
